pyproject.toml
[tool.poetry]
name = "cveuploader"
version = "0.1.0"
description = ""
authors = ["= <tyson@onna.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
click = "^7.1.2"
python-dotenv = "^0.14.0"
pydantic = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.3"
mypy = "^0.782"

[tool.poetry.scripts]
cveuploader="cveuploader.commands:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
